+++ FactoryTest.php
@@ -339,7 +339,7 @@ final class FactoryTest extends TestCase
      */
     public function testCreateFromBaseUri(string $base_uri, string $uri, string $expected): void
     {
-        self::assertSame($expected, Uri::fromClient($uri, $base_uri)->toString());
+        self::assertSame($expected, Uri::fromBaseUri($uri, $base_uri)->toString());
     }
 
     public static function createProvider(): array
@@ -392,20 +392,20 @@ final class FactoryTest extends TestCase
     public function testCreateThrowExceptionWithBaseUriNotAbsolute(): void
     {
         self::expectException(SyntaxError::class);
-        Uri::fromClient('/path/to/you', '//example.com');
+        Uri::fromBaseUri('/path/to/you', '//example.com');
     }
 
     public function testCreateThrowExceptionWithUriNotAbsolute(): void
     {
         self::expectException(SyntaxError::class);
-        Uri::fromClient('/path/to/you');
+        Uri::fromBaseUri('/path/to/you');
     }
 
     public function testCreateWithUriWithoutAuthority(): void
     {
         self::assertSame(
             'data:text/plain;charset=us-ascii,',
-            Uri::fromClient('data:text/plain;charset=us-ascii,')->toString()
+            Uri::fromBaseUri('data:text/plain;charset=us-ascii,')->toString()
         );
     }
 
@@ -413,7 +413,7 @@ final class FactoryTest extends TestCase
     {
         self::assertSame(
             'scheme://host/sky?q#f',
-            Uri::fromClient('scheme://host/path/../sky?q#f')->toString()
+            Uri::fromBaseUri('scheme://host/path/../sky?q#f')->toString()
         );
     }