+++ UriInfo.php
@@ -58,7 +58,7 @@ final class UriInfo
 
         $path = $uri->getPath();
         if ('/' === ($path[0] ?? '') || '' !== $uri->getScheme().$uri->getAuthority()) {
-            $path = UriResolver::resolve($uri, $uri->withPath('')->withQuery($null))->getPath();
+            $path = BaseUri::new($uri->withPath('')->withQuery($null))->resolve($uri)->getPath();
         }
 
         $query = $uri->getQuery();