+++ CHANGELOG.md
@@ -23,19 +23,21 @@ All Notable changes to `League\Uri` will be documented in this file
 - `League\Uri\UriTemplate\Template::expandOrFail`
 - `League\Uri\UriString::parseAuthority`
 - `League\Uri\UriString::buildAuthority`
+- `League\Uri\BaseUri`
 
 ### Fixed
 
-- `League\Uri\UriResolver` and `League\Uri\UriInfo` uri input now supports `Stringable` and `string` type.
+- `League\Uri\UriInfo` uri input now supports `Stringable` and `string` type.
 - `League\Uri\UriTemplate\VariableBag` implements the `IteratorAggregate` interface
 - `League\Uri\UriTemplate\Operator` to improve internal representation when using UriTemplate features.
 
 ### Deprecated
 
+- `League\Uri\UriResolver` use `League\Uri\BaseUri` instead
 - `League\Uri\Uri::createFromString` use `League\Uri\Uri::new`
 - `League\Uri\Uri::createFromUri` use `League\Uri\Uri::new`
 - `League\Uri\Uri::createFromComponents` use `League\Uri\Uri::fromComponents`
-- `League\Uri\Uri::createFromBaseUri` use `League\Uri\Uri::fromClient`
+- `League\Uri\Uri::createFromBaseUri` use `League\Uri\Uri::fromBaseUri`
 - `League\Uri\Uri::createFromServer` use `League\Uri\Uri::fromServer`
 - `League\Uri\Uri::createFromWindowsPath` use `League\Uri\Uri::fromWindowsPath`
 - `League\Uri\Uri::createFromUnixPath` use `League\Uri\Uri::fromUnixPath`
@@ -43,7 +45,7 @@ All Notable changes to `League\Uri` will be documented in this file
 - `League\Uri\Http::createFromString` use `League\Uri\Http::new`
 - `League\Uri\Http::createFromUri` use `League\Uri\Http::new`
 - `League\Uri\Http::createFromComponents` use `League\Uri\Http::fromComponents`
-- `League\Uri\Http::createFromBaseUri` use `League\Uri\Http::fromClient`
+- `League\Uri\Http::createFromBaseUri` use `League\Uri\Http::fromBaseUri`
 - `League\Uri\Http::createFromServer` use `League\Uri\Http::fromServer`
 - `League\Uri\UriTemplate\Template::createFromString` use `League\Uri\UriTemplate\Template::new`